Organise querying into submodule
