+++ src/lib.rs
@@ -11,6 +11,5 @@ mod outpack_file;
 mod hash;
 mod utils;
 
-extern crate pest;
 #[macro_use]
 extern crate pest_derive;
